Reintroduce component for sign-in.
